% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sequencing_qc.R
\name{format_multiqc_stats}
\alias{format_multiqc_stats}
\title{Read and concatenate MultiQC general stats files}
\usage{
format_multiqc_stats(recent_files, rename_map = NULL)
}
\arguments{
\item{recent_files}{A character vector of file paths from \code{get_recent_files}}

\item{rename_map}{A named character vector for renaming samples}
}
\description{
Read and concatenate MultiQC general stats files
}
