% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metrics.R
\name{get_plate_median_metrics}
\alias{get_plate_median_metrics}
\title{Compute median metrics by control groups}
\usage{
get_plate_median_metrics(
  metadata = NULL,
  metrics = "nFeature_RNA",
  group_by = "Treatment_1",
  controls = c("^DMSO|^Media|^Staurosporine|^Untreated"),
  Model_type = NULL
)
}
\arguments{
\item{metadata}{The metadata data frame with nCount_RNA, nFeature_RNA, orig.ident columns, and metadata information.
This metadata should be able to get from Seurat object.}

\item{metrics}{The metrics to compute median for. It can be either "nFeature_RNA" or "nCount_RNA".}

\item{group_by}{The column name in metadata to group by. Default is "Treatment_1".}

\item{controls}{A character vector of control group names to filter by.
Default is c("^DMSO|^Media|^Staurosporine|^Untreated").}

\item{Model_type}{The model type of the plate. It can be either "2D_non_adherent", "2D_adherent", or "3D".}
}
\value{
A data frame with median metrics for each control group for entire plate.
}
\description{
Compute median metrics by control groups
}
