% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sequencing_qc.R
\name{get_recent_files}
\alias{get_recent_files}
\title{Get a list of files matching a pattern in a directory}
\usage{
get_recent_files(
  path,
  pattern = "multiqc_general_stats.txt",
  date = Sys.time()
)
}
\arguments{
\item{path}{The directory to search in}

\item{pattern}{The pattern to match files against}

\item{date}{The date to filter files by modification time,
only getting files with mtime > date}
}
\value{
A character vector of file paths that match the pattern and are modified after the specified date
}
\description{
Get a list of files matching a pattern in a directory
}
\examples{
path <- "/Volumes/bioinf/team_folders/MGC_VCFG/macseq/results"
pattern <- "multiqc_general_stats.txt"
date <- "2025-03-07 00:00:00"
recent_files <- get_recent_files(path = path,
                            pattern = pattern,
                             date = date)
}
