% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mac}
\alias{mac}
\title{Example Seurat objet for a single 384-well plate}
\format{
A Seurat object
}
\source{
From the PMMSq033 plate
}
\usage{
mac
}
\description{
A real Seurat object (stored in \code{data/mac.rda}) used in tests and vignettes.
}
\keyword{datasets}
