% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_qc_metrics.R
\name{read_metrics}
\alias{read_metrics}
\title{Read in CSV files from each plate or}
\usage{
read_metrics(
  path,
  pattern = "*median_genes.csv",
  date = Sys.time(),
  plate = TRUE
)
}
\arguments{
\item{path}{The directory path where the CSV files are located.}

\item{pattern}{The pattern to match files against, default is "*median_genes.csv".
It can be only one of "\*median_genes.csv" or "\*median_counts.csv".}

\item{date}{The date to filter files by modification time, default is the current system time.}

\item{plate}{Logical value indicating whether to read plate-level data (TRUE) or cell type-level data (FALSE).}
}
\value{
A data frame containing the concatenated data from the CSV files.
}
\description{
Read in CSV files from each plate or
}
